Add toy simulation of single-PE peak finding
This macro simulates some number of events of a 100 time samples long channel readout, with a noisy pedestal, and some pulse injected. It explores pedestal subtraction with two definitions of event pedestal. 

Parameters are given as an ordered list of arguments and it runs with ROOT (a standalone C++ or python version would be welcome). They are: number of simulated events, noise level, average injected pulse multiplicity, (average) pulse amplitude, relative width of noise amplitude gaussian.

A few things to explore would be:
- determine pedestal from a histogram, understand biases and if they can be corrected from measured values 
- subtract a fixed pedestal in every event ("truth" or one calculated in a previous run of the simulation)
- quantify which signal/noise ratio is needed for resolving single PE peaks