Add beam trajectory finding macro
This macro uses some previously established electron beam gun position points and angles of incidence (at 4 GeV) to find the bending radius of the beam, and spit out the values of x and theta (angle of incidence) at a given z position along the trajectory. It also makes a plot for visual inspection. 

This macro could be expanded to use the beam energy as an input such that the trajectory can be rescaled accordingly.